requirements.txt
Flask==3.0.0
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
psycopg2-binary==2.9.9
pytz
pandas
openpyxl>=3.1.2
boto3>=1.34.0
openai>=1.0.0
